ocr-pipeline/surya-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
python-doctr[torch]==0.9.0
pdf2image==1.16.3
Pillow==10.2.0
